Added WCF Calling project
Added a separate BizTalk project handling WCF calls to the storage
service. This is because for some reason there is a problem with
adding the service references to the previous BizTalk project that
houses the EDI schemas, which makes that project impossible to compile.
Splitting it out like this has no effect on the BizTalk application, it
just means that the parts required for the integration are deployed in
two different assemblies.